+++ src/pages/home/HomePage.module.css
@@ -60,30 +60,6 @@
   gap: var(--space-lg);
 }
 
-.primaryButton,
-.secondaryButton {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  min-height: var(--button-min-height);
-  border-radius: var(--radius-sm);
-  padding: var(--space-lg) var(--space-xl);
-  font: inherit;
-  text-decoration: none;
-}
-
-.primaryButton {
-  border: none;
-  color: var(--color-primary-contrast);
-  background: var(--gradient-primary);
-}
-
-.secondaryButton {
-  border: var(--border-width) solid var(--color-border-strong);
-  color: var(--color-page);
-  background: var(--color-surface-strong);
-}
-
 .tile {
   min-height: 180px;
 }
