+++ src/pages/recipes/recipes.query.ts
@@ -1,0 +1,20 @@
+import { queryOptions } from '@tanstack/react-query';
+import type { RecipesQueryInput } from './recipes.data';
+import { getRecipes } from './recipes.api';
+
+export const DEFAULT_RECIPES_QUERY_INPUT: RecipesQueryInput = {
+  search: '',
+  nutritionField: 'none',
+  nutritionDirection: 'lte',
+  nutritionValue: null,
+  ratingDirection: 'gte',
+  ratingValue: null,
+  userId: 'mock-user',
+};
+
+export function recipesQueryOptions(input: RecipesQueryInput = DEFAULT_RECIPES_QUERY_INPUT) {
+  return queryOptions({
+    queryKey: ['recipes', input],
+    queryFn: () => getRecipes({ data: input }),
+  });
+}
