+++ src/pages/weight/WeightPage.module.css
@@ -64,19 +64,6 @@
   gap: var(--space-xl);
 }
 
-.primaryButton {
-  min-height: var(--button-min-height);
-  border-radius: var(--radius-pill);
-  padding: 0.65rem 1rem;
-  font: inherit;
-}
-
-.primaryButton {
-  border-color: transparent;
-  background: var(--gradient-primary);
-  color: var(--color-primary-contrast);
-}
-
 .captureField {
   display: grid;
   gap: var(--space-md);
@@ -129,11 +116,6 @@
   width: 1.8rem;
   height: 1.8rem;
   color: var(--color-accent-soft);
-}
-
-.primaryButton {
-  border: none;
-  cursor: pointer;
 }
 
 .logRow {
