+++ src/router.tsx
@@ -1,8 +1,8 @@
 import { QueryClient } from '@tanstack/react-query';
 import { createRouter } from '@tanstack/react-router';
 import { setupRouterSsrQueryIntegration } from '@tanstack/react-router-ssr-query';
-import { DefaultCatchBoundary } from '@/components/DefaultCatchBoundary';
-import { NotFound } from '@/components/NotFound';
+import { DefaultCatchBoundary } from '@/components/default-catch-boundary/DefaultCatchBoundary';
+import { NotFound } from '@/components/not-found/NotFound';
 import { routeTree } from './routeTree.gen';
 
 export function getRouter() {
