+++ src/routes/recipes/add.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { Card } from '@/components/Card';
+import { RecipePlaceholderPage } from '@/pages/recipes/RecipePlaceholderPage';
 
 export const Route = createFileRoute('/recipes/add')({
   component: RecipesAddPage,
@@ -7,11 +7,9 @@ export const Route = createFileRoute('/recipes/add')({
 
 function RecipesAddPage() {
   return (
-    <Card as='article'>
-      <h1>Add recipe</h1>
-      <p>
-        Recipe creation is the next slice. This route exists so the list page can link there now.
-      </p>
-    </Card>
+    <RecipePlaceholderPage
+      body='Recipe creation is the next slice. This route exists so the list page can link there now.'
+      title='Add recipe'
+    />
   );
 }
