+++ src/routes/recipes/view.$id.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { Card } from '@/components/Card';
+import { RecipePlaceholderPage } from '@/pages/recipes/RecipePlaceholderPage';
 
 export const Route = createFileRoute('/recipes/view/$id')({
   component: RecipeViewPage,
@@ -9,10 +9,10 @@ function RecipeViewPage() {
   const { id } = Route.useParams();
 
   return (
-    <Card as='article'>
-      <h1>Recipe view</h1>
-      <p>Recipe detail page is the next slice.</p>
-      <p>Selected recipe id: {id}</p>
-    </Card>
+    <RecipePlaceholderPage
+      body='Recipe detail page is the next slice.'
+      subtitle={`Selected recipe id: ${id}`}
+      title='Recipe view'
+    />
   );
 }
